Moved some classes around to better separate and illustrate the core domain and the api built on top of it.
