+++ src/main/java/com/sethhaskellcondie/thegamepensiveapi/domain/EntityGatewayAbstract.java
@@ -1,7 +1,7 @@
 package com.sethhaskellcondie.thegamepensiveapi.domain;
 
 import com.sethhaskellcondie.thegamepensiveapi.domain.filter.FilterRequestDto;
-import com.sethhaskellcondie.thegamepensiveapi.exceptions.ExceptionResourceNotFound;
+import com.sethhaskellcondie.thegamepensiveapi.domain.exceptions.ExceptionResourceNotFound;
 
 import java.util.List;
