+++ src/main/java/com/sethhaskellcondie/thegamepensiveapi/domain/exceptions/ExceptionInputValidation.java
@@ -1,4 +1,4 @@
-package com.sethhaskellcondie.thegamepensiveapi.exceptions;
+package com.sethhaskellcondie.thegamepensiveapi.domain.exceptions;
 
 import com.sethhaskellcondie.thegamepensiveapi.api.Api;
