+++ src/main/java/com/sethhaskellcondie/thegamepensiveapi/domain/filter/Filter.java
@@ -80,4 +80,3 @@ public class Filter {
     }
 }
 
-record FilterResponseDto(String type, Map<String, String> fields, Map<String, List<String>> filters) { }
