+++ src/test/java/com/sethhaskellcondie/thegamepensiveapi/domain/customfield/CustomFieldValueRepositoryTests.java
@@ -4,8 +4,8 @@ import com.sethhaskellcondie.thegamepensiveapi.domain.Keychain;
 import com.sethhaskellcondie.thegamepensiveapi.domain.filter.Filter;
 import com.sethhaskellcondie.thegamepensiveapi.domain.toy.Toy;
 import com.sethhaskellcondie.thegamepensiveapi.domain.toy.ToyRepository;
-import com.sethhaskellcondie.thegamepensiveapi.exceptions.ExceptionCustomFieldValue;
-import com.sethhaskellcondie.thegamepensiveapi.exceptions.ExceptionMalformedEntity;
+import com.sethhaskellcondie.thegamepensiveapi.domain.exceptions.ExceptionCustomFieldValue;
+import com.sethhaskellcondie.thegamepensiveapi.domain.exceptions.ExceptionMalformedEntity;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
